# back_end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MusicApp API",
        "description": "Documentation de l'API MusicApp (auth, artistes, albums, chansons.)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/chansons": {
            "get": {
                "tags": [
                    "Chansons"
                ],
                "summary": "Get all chansons",
                "description": "Retrieve a list of all songs",
                "operationId": "7188cf1966e57bfbfa559eb745dea39b",
                "responses": {
                    "200": {
                        "description": "List of songs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "titre": {
                                                "type": "string",
                                                "example": "Song 1"
                                            },
                                            "duree": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 3.5
                                            },
                                            "album_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Chansons"
                ],
                "summary": "Create a new chanson",
                "description": "Store a newly created song",
                "operationId": "7bffeda8018231d4cd1a8d125e4e08aa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titre",
                                    "duree",
                                    "album_id"
                                ],
                                "properties": {
                                    "titre": {
                                        "type": "string",
                                        "example": "Song 1"
                                    },
                                    "duree": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3.5
                                    },
                                    "album_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Chanson created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Chanson created successfully"
                                        },
                                        "chanson": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/chansons/{id}": {
            "get": {
                "tags": [
                    "Chansons"
                ],
                "summary": "Get a specific chanson",
                "description": "Retrieve details of a specific song by ID",
                "operationId": "8aa48a3d00b5f59443580da1db2da832",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Chanson ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Song details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "titre": {
                                            "type": "string",
                                            "example": "Song 1"
                                        },
                                        "duree": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 3.5
                                        },
                                        "album_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chanson not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Chansons"
                ],
                "summary": "Update a chanson",
                "description": "Update an existing song",
                "operationId": "4a8684f678449ae0b7d03b7542a2dd5f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titre": {
                                        "type": "string",
                                        "example": "Updated Song"
                                    },
                                    "duree": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4
                                    },
                                    "album_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chanson updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Chanson updated successfully"
                                        },
                                        "chanson": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chanson not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Chansons"
                ],
                "summary": "Delete a chanson",
                "description": "Delete an existing song",
                "operationId": "e7e0069db40b0fbe53db2ab0ea0d2f52",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chanson deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Chanson deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chanson not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/albums/{id}/chansons": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Get all chansons of a specific album",
                "description": "Retrieve all songs that belong to a specific album",
                "operationId": "cf70b36aded058ca7cc18609f2f3486f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Album ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of chansons for the album",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "titre": {
                                                "type": "string",
                                                "example": "Ma Première Chanson"
                                            },
                                            "duree": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 3.5
                                            },
                                            "album_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Album not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use your API token from /login",
                "bearerFormat": "API Token",
                "scheme": "bearer"
            }
        }
    },
...
}